requirements.txt
flask==3.0.0
langchain==0.2.16
langchain-core==0.2.38
langchain-google-genai==1.0.10
langchain-community==0.2.16
langchain-text-splitters==0.2.4
google-generativeai==0.8.3
faiss-cpu==1.7.4
youtube-transcript-api==1.2.4
python-dotenv==1.0.1
tiktoken==0.7.0
requests==2.31.0
